+++ tests/data_context/datasource/test_data_context_datasource_non_sql_methods.py
@@ -102,7 +102,7 @@ def test_get_batch_list_from_new_style_datasource_with_file_system_datasource_co
     os.makedirs(titanic_asset_base_directory_path)
 
     titanic_csv_source_file_path: str = file_relative_path(
-        __file__, "../test_sets/Titanic.csv"
+        __file__, "../../test_sets/Titanic.csv"
     )
     titanic_csv_destination_file_path: str = str(
         os.path.join(base_directory, "data/Titanic_19120414_1313.csv")
